+++ src/plonemeeting/portal/core/tests/test_utils.py
@@ -67,6 +67,7 @@ class TestUtils(PmPortalDemoFunctionalTestCase):
             "&include_all=false"
             "&metadata_fields=file"
             "&metadata_fields=content_category"
+            "&additional_values=publishable"
             "&additional_values=category_title"
             "&additional_values=subcategory_title",
             url,